pyproject.toml
[project]
name = "dspy-rag-intelligence"
version = "0.1.0"
description = "A self-optimizing RAG system with DSPy and Milvus"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dspy-ai>=2.4.0",
    "pymilvus>=2.3.0",
    "sentence-transformers>=2.2.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "streamlit>=1.28.0",
    "watchdog>=3.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "baml-py>=0.215.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
]
